src/requirements.txt
selenium==4.0.0a3
pytest
pytest-html
PyHamcrest
webdriver_manager